+++ src/containerapp/HISTORY.rst
@@ -4,6 +4,7 @@ Release History
 ===============
 upcoming
 ++++++
+* 'az containerapp env --environment-mode': Add environment mode to create and update commands
 
 1.3.0b2
 ++++++